add integrations add protokol bug fixes

+++ core/src/main/java/kz/bsbnb/controller/impl/ReestrControllerImpl.java
@@ -7,12 +7,19 @@ import kz.bsbnb.common.model.*;
 import kz.bsbnb.controller.IReestrController;
 import kz.bsbnb.repository.*;
 import kz.bsbnb.util.CheckUtil;
+import kz.bsbnb.util.ERCBService.*;
 import kz.bsbnb.util.SimpleResponse;
+import org.apache.commons.lang3.time.DateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import javax.xml.rpc.ServiceException;
+import java.rmi.RemoteException;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 
@@ -50,7 +57,7 @@ public class ReestrControllerImpl implements IReestrController {
     public SimpleResponse newHead(@RequestBody @Valid ReestrHead reestrHead) {
         Organisation organisation = organisationRepository.findByOrganisationNum(reestrHead.getIin());
         if (organisation != null) {
-            reestrHead.setDateCreate(new Date());
+            reestrHead.setDateCreate(reestrHead.getDateCreate() == null ? new Date() : reestrHead.getDateCreate());
             reestrHead.setStatus("NEW");
             reestrHead.setOrgName(reestrHead.getOrgName() == null ? organisation.getOrganisationName() : reestrHead.getOrgName());
             reestrHead = reestrHeadRepository.save(reestrHead);
@@ -67,19 +74,23 @@ public class ReestrControllerImpl implements IReestrController {
         if (reestrHead != null) {
             if (!list.isEmpty()) {
                 //Удаляем старые значения
-                for (Reestr reestr : reestrHead.getReestrSet()) {
-                    reestrRepository.deleteByIds(reestr.getId());
+                if (reestrHead.getReestrSet() != null) {
+                    for (Reestr reestr : reestrHead.getReestrSet()) {
+                        reestrRepository.deleteByIds(reestr.getId());
+                    }
                 }
                 int total = list.size();
                 int error = 0;
                 int allShare = 0;
+                String errorText = "";
                 for (Reestr reestr : list) {
                     try {
-                        if (CheckUtil.INN(reestr.getIin())&&CheckUtil.INN(reestr.getVoterIin())) {
+                        if (CheckUtil.INN(reestr.getIin()) && CheckUtil.INN(reestr.getVoterIin())) {
                             reestr.setReestrHeadId(reestrHead);
                             reestrRepository.save(reestr);
                         } else {
                             System.out.println("Error in fill reestr =" + reestr.getVoterIin() + " не прошел проверку");
+                            errorText = errorText + "ИИН/БИН =" + reestr.getVoterIin() + " не прошел проверку/n";
                             error++;
                         }
                     } catch (Exception e) {
@@ -94,7 +105,7 @@ public class ReestrControllerImpl implements IReestrController {
                     reestrHead.setStatus("ERROR");
                 }
                 reestrHeadRepository.save(reestrHead);
-                return new SimpleResponse("Получено - " + total + " записей, из них загружено - " + (total - error) + " записей. Всего голосов загружено акции " + allShare + "!").SUCCESS();
+                return new SimpleResponse(errorText + "Получено - " + total + " записей, из них загружено - " + (total - error) + " записей. Всего голосов загружено акции " + allShare + "!").SUCCESS();
             } else {
                 return new SimpleResponse("Список реестра пуст").ERROR_CUSTOM();
             }
@@ -129,7 +140,7 @@ public class ReestrControllerImpl implements IReestrController {
     @RequestMapping(value = "/headByVotingId/{votingId}", method = RequestMethod.GET)
     public SimpleResponse getHeadList(@PathVariable Long votingId) {
         Voting voting = votingRepository.findOne(votingId);
-        if (voting!=null) {
+        if (voting != null) {
             String iin = voting.getOrganisationId().getOrganisationNum();
             List<ReestrHead> reestrHeads = reestrHeadRepository.findByIin(iin);
             if (!reestrHeads.isEmpty()) {
@@ -160,7 +171,8 @@ public class ReestrControllerImpl implements IReestrController {
                             }
                             //Добавляем новых акционеров
                             List<UserRoles> newRoles = new ArrayList<>();
-                            for (Reestr reestr : reestrHead.getReestrSet()) {
+                            List<Reestr> reestrs = reestrRepository.findByReestrHeadId(reestrHead);
+                            for (Reestr reestr : reestrs) {
                                 User user = userRepository.findByIin(reestr.getIin());
                                 if (user == null) {
                                     user = new User();
@@ -184,6 +196,7 @@ public class ReestrControllerImpl implements IReestrController {
                                     userRoles.setShareCount(reestr.getShareCount());
                                     userRoles.setRole(Role.ROLE_USER);
                                     userRoles.setSharePercent(reestr.getSharePercent());
+                                    userRoles.setShareDate(reestrHead.getDateCreate());
                                     userRoles.setCannotVote(0);
                                     userRoles.setOrgId(voting.getOrganisationId());
                                     userRoles = userRoleRepository.save(userRoles);
@@ -195,7 +208,7 @@ public class ReestrControllerImpl implements IReestrController {
                                 }
                             }
                             //Удаляем голосующих
-                            for (Voter voter:voting.getVoterSet()) {
+                            for (Voter voter : voting.getVoterSet()) {
                                 voterRepository.deleteByIds(voter.getId());
                             }
                             voting.setLastChanged(reestrHead.getDateCreate());
@@ -252,5 +265,115 @@ public class ReestrControllerImpl implements IReestrController {
         }
     }
 
+    @Override
+    @RequestMapping(value = "/checkRegistry/{votingId}", method = RequestMethod.GET)
+    public SimpleResponse checkRegistry(@PathVariable Long votingId, @RequestParam(defaultValue = "01/01/2000") String regDate) {
+        Voting voting = votingRepository.findOne(votingId);
+        if (voting != null) {
+            String bin = voting.getOrganisationId().getOrganisationNum();
+            IERCBVotingServicesservice service = new IERCBVotingServicesserviceLocator();
+            try {
+                IERCBVotingServices iercbVotingServicesPort = service.getIERCBVotingServicesPort();
+                String vDate = regDate.replace("/", ".");
+                int i = iercbVotingServicesPort.existsRegistry(bin, vDate);
+                if (i == 1) {
+                    return new SimpleResponse("Реестр существует").SUCCESS();
+                } else {
+                    return new SimpleResponse("Реестр не найден").ERROR_NOT_FOUND();
+                }
+            } catch (ServiceException e) {
+                return new SimpleResponse("Ошибка при соединении с СВР").ERROR_CUSTOM();
+            } catch (RemoteException e) {
+                return new SimpleResponse("Ошибка при запросе данных").ERROR_CUSTOM();
+            }
+        } else {
+            return new SimpleResponse("Голосование не найдено").ERROR_CUSTOM();
+        }
+    }
+
+    @Override
+    @RequestMapping(value = "/getRegistry/{votingId}", method = RequestMethod.GET)
+    public SimpleResponse getRegistry(@PathVariable Long votingId, @RequestParam(defaultValue = "01/01/2000") String regDate) {
+        Voting voting = votingRepository.findOne(votingId);
+        if (voting != null) {
+            String bin = voting.getOrganisationId().getOrganisationNum();
+            IERCBVotingServicesservice service = new IERCBVotingServicesserviceLocator();
+            try {
+                IERCBVotingServices iercbVotingServicesPort = service.getIERCBVotingServicesPort();
+                String vDate = regDate.replace("/", ".");
+                TResponseRegistry registry = iercbVotingServicesPort.getRegistry(bin, vDate);
+                if (registry.getErrorCode() == 0) {
+                    ReestrHead reestrHead;
+                    SimpleDateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy");
+                    Date truncatedDate = null;
+                    try {
+                        truncatedDate = dateFormat.parse(regDate);
+                    } catch (ParseException e) {
+                        return new SimpleResponse("Неверный формат даты").ERROR_CUSTOM();
+                    }
+                    reestrHead = reestrHeadRepository.findByIinAndDateCreate(bin, truncatedDate);
+                    if (reestrHead == null) {
+                        reestrHead = new ReestrHead();
+                        reestrHead.setDateCreate(truncatedDate);
+                        reestrHead.setIin(bin);
+                        reestrHead.setOrgName(registry.getRegistry().getIssuerName());
+                    } else {
+                        reestrHead.setOrgName(registry.getRegistry().getIssuerName());
+                    }
+                    SimpleResponse result = newHead(reestrHead);
+                    if (result.isSuccess()) {
+                        reestrHead = (ReestrHead) result.getData();
+                        List<Reestr> reestrs = new ArrayList<>();
+                        for (int i = 0; i < registry.getRegistry().getShareholders().length; i++) {
+                            TShareholder shareholder = registry.getRegistry().getShareholders()[i];
+                            Reestr reestr = new Reestr();
+                            reestr.setIin(shareholder.getShareholderIDN());
+                            reestr.setEmail(shareholder.getMail());
+                            reestr.setFirstname(shareholder.getShareholderNoun());
+                            reestr.setName(shareholder.getShareholderSurname() == null || "".equals(shareholder.getShareholderSurname()) ? shareholder.getShareholderName() : shareholder.getShareholderSurname());
+                            reestr.setSurname(shareholder.getShareholderPatronymic());
+                            reestr.setNin(shareholder.getShareholderNIN());
+                            reestr.setAllShareCount((int) shareholder.getAmount());
+                            reestr.setShareCount((int) shareholder.getVoting());
+                            reestr.setPhone(shareholder.getPhone());
+                            reestr.setReestrHeadId(reestrHead);
+                            reestr.setShareType(shareholder.getShareholderKind());
+                            reestr.setSharePercent(shareholder.getPercentVoting());
+                            reestr.setVoterIin(shareholder.getVotingIDN());
+                            reestrs.add(reestr);
+                        }
+                        result = newList(reestrHead.getId(), reestrs);
+                        if (result.isSuccess()) {
+                            result = getList(reestrHead.getId(), votingId);
+                        }
+                    }
+                    return result;
+                } else {
+                    return new SimpleResponse(registry.getErrorText()).ERROR_CUSTOM();
+                }
+            } catch (ServiceException e) {
+                return new SimpleResponse("Ошибка при соединении с СВР").ERROR_CUSTOM();
+            } catch (RemoteException e) {
+                return new SimpleResponse("Ошибка при запросе данных").ERROR_CUSTOM();
+            }
+        } else {
+            return new SimpleResponse("Голосование не найдено").ERROR_CUSTOM();
+        }
+    }
+
+    @Override
+    @RequestMapping(value = "/getChief/{bin}", method = RequestMethod.GET)
+    public SimpleResponse getChief(@PathVariable String bin) {
+        IERCBVotingServicesservice service = new IERCBVotingServicesserviceLocator();
+        try {
+            IERCBVotingServices iercbVotingServicesPort = service.getIERCBVotingServicesPort();
+            String resp = iercbVotingServicesPort.getChief(bin);
+            return new SimpleResponse(resp).SUCCESS();
+        } catch (ServiceException e) {
+            return new SimpleResponse("Ошибка при соединении с СВР").ERROR_CUSTOM();
+        } catch (RemoteException e) {
+            return new SimpleResponse("Ошибка при запросе данных").ERROR_CUSTOM();
+        }
+    }
 
 }
